Fixed: slow gRPC subscriptions