@arwinneil 2020-07-01
``` ini

BenchmarkDotNet=v0.12.0, OS=amzn 2
Intel Xeon Platinum 8259CL CPU 2.50GHz, 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.301
  [Host]     : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT
  Job-YAIOFF : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT

InvocationCount=1  MaxIterationCount=50  UnrollFactor=1  
WarmupCount=20  

``` 
```
|                       Type |                         Method | numberOfEvents | eventSize | batchSize | warmTcp |       Mean |      Error |    StdDev |     Median |
|--------------------------- |------------------------------- |--------------- |---------- |---------- |-------- |-----------:|-----------:|----------:|-----------:|
|            TcpV6Benchmarks |           v6Tcp_v6Server_Write |           1000 |       100 |       100 |   False | 108.705 ms | 13.4982 ms | 27.267 ms | 121.622 ms |
|            TcpV6Benchmarks |            v6Tcp_v6Server_Read |           1000 |       100 |       100 |   False |  14.548 ms |  2.3432 ms |  4.165 ms |  13.140 ms |
|            TcpV6Benchmarks |           v6Tcp_v6Server_Write |           1000 |       100 |       100 |    True |  54.112 ms |  7.0405 ms | 13.565 ms |  53.326 ms |
|            TcpV6Benchmarks |            v6Tcp_v6Server_Read |           1000 |       100 |       100 |    True |         NA |         NA |        NA |         NA |
|            TcpV6Benchmarks |           v6Tcp_v6Server_Write |           1000 |       100 |      1000 |   False |  93.196 ms |  3.7031 ms |  7.480 ms |  94.431 ms |
|            TcpV6Benchmarks |            v6Tcp_v6Server_Read |           1000 |       100 |      1000 |   False |   8.232 ms |  0.8981 ms |  1.596 ms |   7.619 ms |
|            TcpV6Benchmarks |           v6Tcp_v6Server_Write |           1000 |       100 |      1000 |    True |  48.296 ms | 16.2015 ms | 32.728 ms |  46.065 ms |
|            TcpV6Benchmarks |            v6Tcp_v6Server_Read |           1000 |       100 |      1000 |    True |         NA |         NA |        NA |         NA |
|  TcpSubscriptionBenchmarks |  v6Tcp_v6Server_SubscribeToAll |           1000 |         ? |       100 |   False |   5.491 ms |  2.5362 ms |  4.947 ms |   3.988 ms |
|  TcpSubscriptionBenchmarks |  v6Tcp_v6Server_SubscribeToAll |           1000 |         ? |       100 |    True |         NA |         NA |        NA |         NA |
|             GrpcBenchmarks |          v6Grpc_v6Server_Write |           1000 |       100 |         ? |       ? |  28.049 ms |  1.9154 ms |  3.825 ms |  26.590 ms |
|             GrpcBenchmarks |           v6Grpc_v6Server_Read |           1000 |       100 |         ? |       ? |  22.143 ms |  0.7130 ms |  1.424 ms |  21.810 ms |
|            TcpV6Benchmarks |           v6Tcp_v6Server_Write |           1000 |      1000 |       100 |   False |  90.577 ms |  2.1331 ms |  4.210 ms |  90.356 ms |
|            TcpV6Benchmarks |            v6Tcp_v6Server_Read |           1000 |      1000 |       100 |   False |  14.155 ms |  2.1738 ms |  3.920 ms |  12.844 ms |
|            TcpV6Benchmarks |           v6Tcp_v6Server_Write |           1000 |      1000 |       100 |    True |  76.708 ms |  6.1569 ms | 12.437 ms |  79.936 ms |
|            TcpV6Benchmarks |            v6Tcp_v6Server_Read |           1000 |      1000 |       100 |    True |   8.436 ms |  1.5849 ms |  2.898 ms |   7.237 ms |
|            TcpV6Benchmarks |           v6Tcp_v6Server_Write |           1000 |      1000 |      1000 |   False |  62.404 ms |  9.5013 ms | 18.975 ms |  55.065 ms |
|            TcpV6Benchmarks |            v6Tcp_v6Server_Read |           1000 |      1000 |      1000 |   False |         NA |         NA |        NA |         NA |
|            TcpV6Benchmarks |           v6Tcp_v6Server_Write |           1000 |      1000 |      1000 |    True |  60.692 ms | 14.1395 ms | 28.562 ms |  69.837 ms |
|            TcpV6Benchmarks |            v6Tcp_v6Server_Read |           1000 |      1000 |      1000 |    True |   4.807 ms |  0.7531 ms |  1.451 ms |   3.943 ms |
|  TcpSubscriptionBenchmarks |  v6Tcp_v6Server_SubscribeToAll |           1000 |         ? |      1000 |   False |   2.749 ms |  0.8716 ms |  1.658 ms |   2.905 ms |
|  TcpSubscriptionBenchmarks |  v6Tcp_v6Server_SubscribeToAll |           1000 |         ? |      1000 |    True |         NA |         NA |        NA |         NA |
|             GrpcBenchmarks |          v6Grpc_v6Server_Write |           1000 |      1000 |         ? |       ? |  38.983 ms |  2.5297 ms |  5.052 ms |  39.903 ms |
|             GrpcBenchmarks |           v6Grpc_v6Server_Read |           1000 |      1000 |         ? |       ? |  21.537 ms |  0.7783 ms |  1.518 ms |  21.401 ms |
| GrpcSubscriptionBenchmarks | v6Grpc_v6Server_SubscribeToAll |           1000 |         ? |         ? |       ? |  22.260 ms |  1.0636 ms |  2.074 ms |  22.299 ms |

Benchmarks with issues:
  TcpV6Benchmarks.v6Tcp_v6Server_Read: Job-YAIOFF(InvocationCount=1, MaxIterationCount=50, UnrollFactor=1, WarmupCount=20) [numberOfEvents=1000, eventSize=100, batchSize=100, warmTcp=True]
  TcpV6Benchmarks.v6Tcp_v6Server_Read: Job-YAIOFF(InvocationCount=1, MaxIterationCount=50, UnrollFactor=1, WarmupCount=20) [numberOfEvents=1000, eventSize=100, batchSize=1000, warmTcp=True]
  TcpSubscriptionBenchmarks.v6Tcp_v6Server_SubscribeToAll: Job-YAIOFF(InvocationCount=1, MaxIterationCount=50, UnrollFactor=1, WarmupCount=20) [numberOfEvents=1000, batchSize=100, warmTcp=True]
  TcpV6Benchmarks.v6Tcp_v6Server_Read: Job-YAIOFF(InvocationCount=1, MaxIterationCount=50, UnrollFactor=1, WarmupCount=20) [numberOfEvents=1000, eventSize=1000, batchSize=1000, warmTcp=False]
  TcpSubscriptionBenchmarks.v6Tcp_v6Server_SubscribeToAll: Job-YAIOFF(InvocationCount=1, MaxIterationCount=50, UnrollFactor=1, WarmupCount=20) [numberOfEvents=1000, batchSize=1000, warmTcp=True]