@arwinneil 2020-07-07
## Catch Subscription Benchmark
Time from write request being sent to server to event being received by client, average of 100 events.
Amazon Linux 2, m5.2xlarge , 1000 GB SSD with 3000 IOPS

master @ 065d2e143379f5d70c48c516c804bee224cb7eb3

```
 ---------------------------------------------------
 | Benchmark             | Mean        | StdDev    |
 ---------------------------------------------------
 | gRPC SubscribeFromEnd | 3304.397 ms | 54.418 ms |
 ---------------------------------------------------
 | TCP SubscribeFromEnd  | 2.083 ms    | 1.454 ms  |
 ---------------------------------------------------
```

PR #2566
```
 --------------------------------------------------- 
 | Benchmark             | Mean        | StdDev    | 
 --------------------------------------------------- 
 | gRPC SubscribeFromEnd | 3317.896 ms | 58.789 ms | 
 --------------------------------------------------- 
 | TCP SubscribeFromEnd  | 2.207 ms    | 1.610 ms  | 
 --------------------------------------------------- 
```