@thefringeninja 2020-07-07
@arwinneil out of curiosity what happens when you make an initial request (e.g. reading from all) before doing the benchmark? I'm wondering if there is some large amount of magic happening in Kestrel startup (like IIS).