@arwinneil 2020-07-07
@thefringeninja That very first request establishes the TCP connection to the server. That usually takes a little more time and including it in results skews the average.  Benchmarks with warm TCP connections make that first read request, while cold TCP benchmarks don't.